requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
python-dotenv==1.1.0
openpyxl==3.1.5
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langgraph==0.4.8
faiss-cpu==1.11.0
boto3==1.38.38
langchain-aws==0.2.26